Fixes #25293 - Support Puma
To support Puma, this takes the approach of introducing a Rack
application that understands the HTTP request scheme. Depending on that,
it responds either with the HTTP or HTTPS app.

This means all the responsibility of binding to HTTP and HTTPS is
shifted to the application server.

To use this:

  bundle exec puma -b tcp://127.0.0.1:8000 -b 'ssl://127.0.0.1:8443?key=config/key.pem&cert=config/cert.pem'

Note that this doesn't set the secure ciphers nor protocols. It also
requires Puma to be built with OpenSSL extensions to be able to bind on
HTTPS.